kernel: move time_units into sys subdirectory
This isn't something the user will ever include directory, so take
steps to hide it.
